refactor: Rename core module
- view -> producer
- render -> consumer
